requirements.txt
django
django-tailwind
pillow
django-cleanup
djangorestframework
markdown
django-filter
django-cors-headers
recipe-scrapers
openai